vcbackend/requirements.txt
#Tested in Python 3.7, Linux
argcomplete==2.0.0
asgiref==3.6.0
boto3==1.26.71
botocore==1.29.71
certifi==2022.12.7
cfn-flip==1.3.0
charset-normalizer==3.0.1
click==8.1.3
Collectfast==2.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django #==4.1.7
django-storages==1.13.2
djangorestframework==3.14.0
durationpy==0.5
environ==1.0
hjson==3.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.2
openapi-codec==1.3.2
placebo==0.9.0
#psycopg2==2.9.5
psycopg2-binary
python-dateutil==2.8.2
python-dotenv==0.21.1
python-slugify==8.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.1
troposphere==4.3.0
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.2.2
zappa==0.56.1
